doc-search/requirements.txt
# ── UI ─────────────────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Vector DB ───────────────────────────────────────────────────────────────
qdrant-client>=1.9.0

# ── Embedding (로컬 우선) ────────────────────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.2.0

# ── LlamaIndex (오케스트레이션 / 텍스트 청킹) ──────────────────────────────────
llama-index-core>=0.10.0
llama-index-vector-stores-qdrant>=0.2.0
llama-index-embeddings-huggingface>=0.2.0

# ── 하이브리드 검색 (BM25) ────────────────────────────────────────────────────
rank-bm25>=0.2.2
numpy>=1.26.0

# ── 문서 파싱 ─────────────────────────────────────────────────────────────────
PyMuPDF>=1.24.0          # PDF
python-pptx>=0.6.23      # PowerPoint
python-docx>=1.1.0       # Word
openpyxl>=3.1.2          # Excel
pandas>=2.2.0            # Excel (데이터 처리)
olefile>=0.47            # HWP (OLE 컨테이너 파싱)
# pyhwpx               # HWP 고급 추출 (선택적 설치: pip install pyhwpx)

# ── Windows 자동화 ────────────────────────────────────────────────────────────
# pywin32              # Office COM 자동화 (선택적 설치: pip install pywin32)

# ── 유틸리티 ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
